requirements.txt
# Colab / Linux + NVIDIA GPU 推荐用本文件。macOS 本机见 requirements-macos.txt
unsloth>=2025.1.0
trl>=0.14.0
peft>=0.14.0
bitsandbytes>=0.45.0
transformers>=4.49.0
datasets>=3.2.0
accelerate>=1.2.0
torch>=2.5.0
pyyaml>=6.0.2
numpy>=2.0.0
pandas>=2.2.3
matplotlib>=3.10.0
tqdm>=4.67.0
